refactor: 게임 상태를 변경하는 GameService 분리
